% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{calib}
\alias{calib}
\title{Calibration of a data.frame}
\usage{
calib(x, models = list(0), append = T)
}
\arguments{
\item{models}{A list of models}

\item{append=true}{Boolean indicating if the output of each model should be appended to the right of the data.frame
If append=false the operation of each model is on the output of the previous one.}
}
\value{
A calibrated data.frame
}
\description{
Calibration of a data.frame
}
\examples{
  n = 100
  rows = sample(1:nrow(iris), n)
  train = iris[n,]
  sel = 1:30
  calib = train[sel,]
  learn = train[-sel,]
  m1 = conti2bins(calib[,1:4])
  m2 = conti2norm(calib[,1:4])
  M = list()
  M[1] = m1
  M[2] = m2
  learn = calib(learn, M, append=T)

}
