% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{cat2num}
\alias{cat2num}
\title{Converts categorical variables to continuous variables}
\usage{
cat2num(x, y, ...)
}
\arguments{
\item{x}{:  data.frame with independent variables of type category}

\item{y}{: character vector}
}
\value{
cat2num object
}
\description{
Based on the mutual information MI(xi, y) each categorical
variable value is represented by its score
}
\examples{
  m = cat2num()
  predict(m, )

}
