% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{cbc}
\alias{cbc}
\title{Clustering by committee}
\usage{
cbc(x, top = 50, theta1 = 0.35, theta2 = 0.25,
  method = "Euclidean", lim = nrow(x)/2)
}
\arguments{
\item{x}{A data.frame with independent variables of type continous}

\item{top}{A subset of highly similar entries from which the initial cluster is built}

\item{theta1}{Similarity threshold between centroids of clusters}

\item{theta2}{Similarity threshold between an observation and a committee}

\item{method}{Similarity metric available from the "proxy" package:
"Jaccard"
"Kulczynski1" "Kulczynski2"
"Mountford"
"Fager"
"Russel"
"simple matching"
"Hamman"
"Faith"
"Tanimoto"
"Dice"
"Phi"
"Stiles"
"Michael"
"Mozley"
"Yule" "Yule2"
"Ochiai"
"Simpson"
"Braun-Blanquet"
"cosine"
"eJaccard" "fJaccard"
"correlation"
"Chi-squared"
"Phi-squared"
"Tschuprow"
"Cramer"
"Pearson"
"Gower"
"Euclidean"
"Mahalanobis"
"Bhjattacharyya"
"Manhattan"
"supremum"
"Minkowski"
"Canberra"
"Wave"
"divergence"
"Kullback"
"Bray"
"Soergel"
"Levenshtein"
"Podani"
"Chord"
"Geodesic"
"Whittaker"
"Hellinger"}
}
\value{
A model to predict cluster
}
\description{
Clustering by committee
}
\examples{
  n = 100
  rows = sample(1:nrow(iris), n)
  train = iris[n,]
  test = iris[-rows,]
  m = cbc(train[,1:4])
  p = predict(m, test)

}
