% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{const2rm}
\alias{const2rm}
\title{Remove constant variables}
\usage{
const2rm(x)
}
\arguments{
\item{x}{: data.frame with categorical or numerical variables}
}
\value{
const2rm object
}
\description{
Remove constant variables
}
\examples{
  m = const2rm(iris[,1:4])
  predict(m, iris)

}
