% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{cor2rm}
\alias{cor2rm}
\title{Remove correlating variables}
\usage{
cor2rm(x, thresh = 0.95)
}
\arguments{
\item{x}{: data.frame with numeric variables}

\item{thresh}{: DOUBLE;
IF cor > thresh
THEN exclude variable}
}
\value{
cor2rm object
}
\description{
Remove correlating variables
}
\examples{
  m = cor2rm(iris[,1:4])
  predict(m, iris)

}
