% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2dict}
\alias{data2dict}
\title{Data dictionary definition based on input data}
\usage{
data2dict(x = NULL, file = NULL, sep = ",")
}
\arguments{
\item{x}{: data.frame with categorical or numerical variables}

\item{file}{: STRING, e.g. "dict.csv"}

\item{sep}{: CHAR, default ","}
}
\value{
data2dict object

  if file is not null, then a file is created with the columns:-
      var := [x|y|-]
    class := [cat|num]
     name := STRING
}
\description{
Data dictionary definition based on input data
}
\examples{
  dict = data2dict(iris, file="dict.csv")
  ## emacs -nw dict.csv ## edit var=y
  dict = data2dict(file="dict.csv")
  data = predict(dict, iris)

}
