% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2formula}
\alias{data2formula}
\title{Create formula from dict}
\usage{
data2formula(model, newx, varx = NULL, vary = NULL)
}
\arguments{
\item{model}{: data2dict object}

\item{newx}{: data.frame}
}
\value{
formula object
}
\description{
Create formula from dict
}
