% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2isof}
\alias{data2isof}
\title{Anomaly Detection:  isofor: iForest}
\usage{
data2isof(x, ...)
}
\description{
Anomaly Detection:  isofor: iForest
}
\examples{
  m = data2isofor(iris[,1:4])
  p = predict(p, iris)

}
