% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2kernlab}
\alias{data2kernlab}
\title{Support Vector Machines: kernlab::ksvm with rbfdot, polydot, ..., kernels}
\usage{
data2kernlab(f, x)
}
\arguments{
\item{f}{: formula}

\item{x}{: data.frame}
}
\description{
Support Vector Machines: kernlab::ksvm with rbfdot, polydot, ..., kernels
}
\examples{
  ksvm = kernlab::ksvm(Species ~ ., data=iris)
  p = predict(ksvm, newx=iris)
  table(p, a=iris[["Species"]])

}
