% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2nb}
\alias{data2nb}
\title{Naive Bayes:   e1071: naiveBayes}
\usage{
data2nb(f, x, ...)
}
\description{
Naive Bayes:   e1071: naiveBayes
}
\examples{
  iris[["Species"]] = as.factor(iris[["Species"]])
  m = data2nb(Species ~ ., iris)
  predict(m, newdata=iris, type="class")

}
