% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2nnet}
\alias{data2nnet}
\title{Neural Networks:  nnet::nnet}
\usage{
data2nnet(f, x, ...)
}
\description{
Neural Networks:  nnet::nnet
}
\examples{
  m = data2nnet(Species ~ ., iris)
  predict(m, iris)

}
