% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2rf}
\alias{data2rf}
\title{Tree-based models and ensembles:  randomForest::randomForest}
\usage{
data2rf(f, x)
}
\arguments{
\item{f}{: formula}

\item{x}{: data.frame}
}
\description{
Tree-based models and ensembles:  randomForest::randomForest
}
\examples{
  iris[["Species"]] = as.factor(iris[["Species"]])
  dt = randomForest::randomForest(Species ~ ., data=iris)
  p = predict(dt, newdata=iris)
  table(p, a=iris[["Species"]])

}
