% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2sets}
\alias{data2sets}
\title{Create subsets for learning, evaluating or testing}
\usage{
data2sets(y, t = NULL, sample = "down", split = 0.5, scale = 1,
  nrep = 10)
}
\arguments{
\item{y}{:character vector; a label is an element of
a set of length two, e.g. {0, 1}}
}
\value{
list : with
    $id_train := vector/row index
    $id_eval := vector/row index
    $id_learn := vector/row index
    $reps := 1:nrep
    $nrep := INTEGER
}
\description{
Create subsets for learning, evaluating or testing
}
\examples{
  iris$y = ifelse(iris$Species == "", 1, 0)
  sets = data2sets(iris$y, sample="down")
  train = iris[sets$id_train,]
  model = svm(y ~ ., data=train)
  eval = iris[sets$id_eval,]
  predict(model, newx=eval)

  ## learn multiple models
  models = lapply(sets$reps, function(r){
    learn = iris[sets$id_learn_[[r]],]
    svm(y ~ ., data=learn)
  })
  ## average models

}
