% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2svm}
\alias{data2svm}
\title{Support Vector Machines:  e1071::svm}
\usage{
data2svm(f, x)
}
\arguments{
\item{f}{: formula}

\item{x}{: data.frame}
}
\description{
Support Vector Machines:  e1071::svm
}
\examples{
  X = 1:2
  Y = 5
  iris[["Species"]] = as.factor(iris[["Species"]])
  svm = e1071::svm(Species ~ ., data=iris[,c(Y, X)])
  p = predict(svm, newdata=iris[,X])##--important
  table(p, a=iris[["Species"]])

}
