% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2svm1}
\alias{data2svm1}
\title{Anomaly Detection:   e1071: svm with “one-classification” type}
\usage{
data2svm1(f, x, ...)
}
\description{
Anomaly Detection:   e1071: svm with “one-classification” type
}
\examples{
  m = data2svm1(Species ~ ., iris)
  predict(m, iris)

}
