% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2xbg}
\alias{data2xbg}
\title{Tree-based models and ensembles:  xgboost::xgb.Booster  with multi:softprob, multi:softmax, and binary:logistic objectives}
\usage{
data2xbg(f, x, y)
}
\description{
Tree-based models and ensembles:  xgboost::xgb.Booster  with multi:softprob, multi:softmax, and binary:logistic objectives
}
\examples{
  iris[["Species"]] = as.character(iris[["Species"]])
  xgb = data2xgb(Species ~ ., iris, iris[["Species"]]
  p = predict(xgb, iris)
  table(p, a=iris[["Species"]]

}
