% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2xcat}
\alias{data2xcat}
\title{Extract categorical variables on the basis of the current data.frame}
\usage{
data2xcat(x, model)
}
\arguments{
\item{x}{: data.frame}

\item{model}{: dictionary}
}
\value{
vector with categorical variable names
}
\description{
Extract categorical variables on the basis of the current data.frame
}
