% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2xnum}
\alias{data2xnum}
\title{Extract numerical variables on the basis of the current data.frame}
\usage{
data2xnum(x, model)
}
\arguments{
\item{x}{: data.frame}

\item{model}{: dictionary}
}
\value{
vector with numerical variable names
}
\description{
Extract numerical variables on the basis of the current data.frame
}
