% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{dict2x}
\alias{dict2x}
\title{Extract x variables}
\usage{
dict2x(model, newx, ...)
}
\arguments{
\item{model}{: data2dict object}

\item{newx}{: data.frame}
}
\value{
character vector
}
\description{
Extract x variables
}
