% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{dict2xcat}
\alias{dict2xcat}
\title{Extract categorical x variables}
\usage{
dict2xcat(model, newx, ...)
}
\arguments{
\item{model}{: data2dict object}

\item{newx}{: data.frame}
}
\value{
character vector
}
\description{
Extract categorical x variables
}
