% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{dict2xnum}
\alias{dict2xnum}
\title{Extract numeric x variables}
\usage{
dict2xnum(model, newx, ...)
}
\arguments{
\item{model}{: data2dict object}

\item{newx}{: data.frame}
}
\value{
character vector
}
\description{
Extract numeric x variables
}
