% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{dict2y}
\alias{dict2y}
\title{Extract y variables}
\usage{
dict2y(model, newx, ...)
}
\arguments{
\item{model}{: data2dict object}

\item{newx}{: data.frame}
}
\value{
character vector
}
\description{
Extract y variables
}
