% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{na2rm}
\alias{na2rm}
\title{Remove NA variables}
\usage{
na2rm(x, thresh = 0.05)
}
\arguments{
\item{x}{: data.frame with categorical or numerical variables}

\item{thresh}{: DOUBLE
IF num_NA / len_var > thresh
THEN remove}
}
\value{
na2rm object
}
\description{
Remove NA variables
}
\examples{
  m = na2rm(iris[,1:4])b
  predict(m, iris)

}
