% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{nbtree}
\alias{nbtree}
\title{NBTree}
\usage{
nbtree(x, y, loss = matrix(c(0, 1, 5, 0), ncol = 2, byrow = TRUE))
}
\arguments{
\item{x}{: data.frame with numeric variables}

\item{y}{: character; class label}
}
\value{
nbtree object
}
\description{
Scaling up the accuracy of NB classifiers: a decision tree hybrid -- Ron Kohavi
}
\examples{
  tmp = iris[1:100,]
  tmp[["Species"]] = ifelse(tmp[["Species"]] == "setosa", 1, 0)
  m = nbtree(tmp[,1:4], tmp[["Species"]])
  p = predict(m, iris)

}
