% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2bcp}
\alias{num2bcp}
\title{Box/Cox Power (BCP) Transform}
\usage{
num2bcp(x, ...)
}
\arguments{
\item{x}{: data.frame}
}
\value{
num2bcp object
}
\description{
Box/Cox Power (BCP) Transform
}
\examples{
  m = num2bcp(iris[,1:2])
  predict(m, iris[,1:4])

}
