% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2bin}
\alias{num2bin}
\title{Converts a continuous variable to bins}
\usage{
num2bin(x, ...)
}
\arguments{
\item{x}{: ata.frame with independent variables of type continous}
}
\value{
num2bin object
}
\description{
Converts a continuous variable to bins
}
\examples{
  n = 100
  rows = sample(1:nrow(iris), n)
  train = iris[n,]
  sel = 1:30
  calib = train[sel,]
  learn = train[-sel,]
  m = num2bin(calib[,1:4])
  learn = predict(m, learn)

}
