% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2div}
\alias{num2div}
\title{Feature interaction:  f1 / f2}
\usage{
num2div(x, y = NULL, thresh = 0.5, ntree = 10, smooth = 0.001)
}
\arguments{
\item{x}{: data.frame with numerical variables}

\item{y}{: vector with class label}

\item{thresh}{: numerical;  selects variables with importance >= thresh}

\item{ntree}{: integer;  number of trees}
}
\value{
num2div object
}
\description{
Feature interaction:  f1 / f2
}
\examples{
  df = iris[1:100,]
  df$Species = as.factor(ifelse(df$Species == "setosa", 1, 0))
  m = num2div(df[,1:4], df[,5])
  xdiv = predict(m, df[,1:4]
  data = cbind(df, xdiv)

}
