% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2dtf}
\alias{num2dtf}
\title{Decision tree feature}
\usage{
num2dtf(f, x, loss = matrix(c(0, 1, 5, 0), ncol = 2, byrow = TRUE))
}
\arguments{
\item{f}{: formula}

\item{x}{: data.frame with numeric variables and class label}

\item{loss}{: matrix; describing the loss:
c(TP, FP,
  FN, TN)}
}
\value{
num2dtf object
}
\description{
Decision tree feature
}
\examples{
  m = num2dtf(Species ~ ., iris)
  dtf = predict(m, iris)
  cbind(iris, dtf)

}
