% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2imp}
\alias{num2imp}
\title{Feature selection using randomForest}
\usage{
num2imp(x, y, thresh = 0.5, ntree = 10)
}
\arguments{
\item{x}{: data.frame with numerical variables and class label}

\item{thresh}{: numerical;  selects variables with importance >= thresh}

\item{ntree}{: integer;  number of trees}

\item{f}{: formula}
}
\value{
num2imp object
}
\description{
Features are selected on the basis of variable$importance
}
\examples{
  m = num2imp(Species ~ ., iris)
  dat = predict(m, iris)

}
