% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2int}
\alias{num2int}
\title{Interaction of numerical variables}
\usage{
num2int(x, t = NULL)
}
\arguments{
\item{x}{: data.frame with numerical variables
if t=NULL (default), then all variable combinations are used and four
different interaction operations used to compute new variables:
a*b (new variable infix = name(a) _dot_ name(b)
a+b (new variable infix = name(a) _sum_ name(b)
a-b (new variable infix = name(a) _dif_ name(b)
a/b (new variable infix = name(a) _div_ name(b)}

\item{t}{: numeric vector;  default NULL
if available then each numeric variable is multiplied by t: t*xi}
}
\value{
data.frame
}
\description{
Interaction of numerical variables
}
\examples{
  num2int(iris[,1:3])

}
