% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2log}
\alias{num2log}
\title{Log numeric variables}
\usage{
num2log(x, smooth = 0.001)
}
\arguments{
\item{x}{: data.frame; with numeric variables}

\item{smooth}{: numeric; used to add to x before log:  log( x+smooth )}
}
\value{
data.frame; number of dimensions may be reduced
}
\description{
Log numeric variables
}
\examples{
  xlog = num2log(iris[,1:4])

}
