% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2norm}
\alias{num2norm}
\title{Continuous variable scaling}
\usage{
num2norm(x, method = "minmax", ...)
}
\arguments{
\item{x}{: data.frame with independent variables of type continous}

\item{method}{: string = ["minmax", "znorm"]}
}
\value{
num2norm object
}
\description{
Continuous variable scaling
}
\examples{
  m = num2norm(iris[,1:2])
  predict(m, iris[,1:4])

}
