% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2pca}
\alias{num2pca}
\title{PCA based transformation of numerical data}
\usage{
num2pca(x, ...)
}
\arguments{
\item{x}{: data.frame with numerical variables}
}
\value{
num2pca object
}
\description{
PCA based transformation of numerical data
}
\examples{
  m = num2pca(iris[,1:4])
  predict(m, iris)  ##--TODO

}
