% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2profile}
\alias{num2profile}
\title{Extract profile information from numerical variables}
\usage{
num2profile(x, group, width = 100, lag = as.integer(width/2))
}
\arguments{
\item{x}{: data.frame with numerical variables}

\item{group}{: a vector with group variable}

\item{width}{: INTEGER,  window for rollapply operation}

\item{lag}{: INTEGER,  used to compute the normalised difference index
of n and r = lag(n, k)}
}
\value{
a data.frame with
  z
  min
  max
  median
  ndi
}
\description{
Extract profile information from numerical variables
}
\examples{
  df = num2profile(iris[,1:4], iris[["Species"]], width=10)
  learn = cbind(iris[["Species"]], df)
  svm = e1071::svm(Species ~ ., data=learn)

}
