% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2rf}
\alias{num2rf}
\title{Feature selection using randomForest}
\usage{
num2rf(f, x, thresh = 5, ntree = 10)
}
\arguments{
\item{f}{: formula}

\item{x}{: data.frame with numerical variables and class label}

\item{thresh}{: numerical;  selects variables with importance >= thresh}

\item{ntree}{: integer;  number of trees}
}
\value{
num2rf object
}
\description{
Features are selected on the basis of variable$importance
}
\examples{
  m = num2rf(Species ~ ., iris)
  dat = predict(m, iris)

}
