% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2sq}
\alias{num2sq}
\title{Square numeric variables}
\usage{
num2sq(x)
}
\arguments{
\item{x}{: data.frame; with numeric variables}
}
\value{
data.frame; number of dimensions may be reduced
}
\description{
Square numeric variables
}
\examples{
  xsq = num2sq(iris[,1:4])

}
