% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{num2trafo}
\alias{num2trafo}
\title{Box-Cox transformation}
\usage{
num2trafo(x, ...)
}
\arguments{
\item{x}{: data.frame}
}
\value{
num2trafo object
}
\description{
Box-Cox transformation
}
\examples{
  m = num2trafo(iris[,1:2])
  predict(m, iris[,1:4])

}
