% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{nzv2rm}
\alias{nzv2rm}
\title{Remove near zero variance (NZV) variables}
\usage{
nzv2rm(x, ratio = 0.5, thresh = 0.05)
}
\arguments{
\item{x}{: data.frame with numerical variables}

\item{ratio}{: DOUBLE;
IF num_NA / len_var >= RATIO
THEN  remove NA}

\item{thresh}{: DOUBLE;
IF nzv < thresh
THEN  exclude NZV variables}
}
\value{
nzv2rm object
}
\description{
Remove near zero variance (NZV) variables
}
\examples{
  sel = rm_nzv(iris[,1:4])
  learn = iris[,sel]

}
