% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.cat2num}
\alias{predict.cat2num}
\title{Converts categorical variables to continuous variables}
\usage{
\method{predict}{cat2num}(model, newx, ...)
}
\arguments{
\item{model}{: cat2num object}

\item{newx}{: data.frame}
}
\value{
data.frame;  dim is preserved
}
\description{
Converts categorical variables to continuous variables
}
\examples{
  m = cat2num()
  predict(m, )

}
