% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.const2rm}
\alias{predict.const2rm}
\title{Remove constant variables}
\usage{
\method{predict}{const2rm}(model, newx)
}
\arguments{
\item{model}{: const2rm object}

\item{newx}{: data.frame}
}
\value{
data.frame;  number of columns is reduced
}
\description{
Remove constant variables
}
\examples{
  m = const2rm(iris[,1:4])
  predict(m, iris)

}
