% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.cor2rm}
\alias{predict.cor2rm}
\title{Remove correlating variables}
\usage{
\method{predict}{cor2rm}(model, newx)
}
\arguments{
\item{model}{: cor2rm object}

\item{newx}{: data.frame}
}
\value{
data.frame : number of columns is reduced
}
\description{
Remove correlating variables
}
\examples{
  m = cor2rm(iris[,1:4])
  predict(m, iris)

}
