% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.data2dict}
\alias{predict.data2dict}
\title{Data dictionary applied on input data}
\usage{
\method{predict}{data2dict}(model, newx, ...)
}
\arguments{
\item{model}{: data2dict object}

\item{newx}{: data.frame}
}
\value{
data.frame;  dim is preserved
}
\description{
Operations include:-
  - casting x numeric variables
  - casting x categorical variables
  - casting y variable [TODO]
}
\examples{
  dict = data2dict(iris, file="dict.csv")
  ## emacs -nw dict.csv ## edit var=y
  dict = data2dict(file="dict.csv")
  data = predict(dict, iris)

}
