% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.na2rm}
\alias{predict.na2rm}
\title{Remove NA variables}
\usage{
\method{predict}{na2rm}(model, newx)
}
\arguments{
\item{model}{: na2rm object}

\item{newx}{: data.frame}
}
\value{
data.frame;  number of columns is reduced
}
\description{
Remove NA variables
}
\examples{
  m = na2rm(iris[,1:4])b
  predict(m, iris)

}
