% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2bin}
\alias{predict.num2bin}
\title{Converts a continuous variable to bins}
\usage{
\method{predict}{num2bin}(model, newx, ...)
}
\arguments{
\item{model}{: num2bin object}

\item{newx}{: data.frame}
}
\value{
data.frame;  dim is preserved
}
\description{
Converts a continuous variable to bins
}
\examples{
  n = 100
  rows = sample(1:nrow(iris), n)
  train = iris[n,]
  sel = 1:30
  calib = train[sel,]
  learn = train[-sel,]
  m = num2bin(calib[,1:4])
  learn = predict(m, learn)

}
