% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2dif}
\alias{predict.num2dif}
\title{Feature interaction:  f1 - f2}
\usage{
\method{predict}{num2dif}(model, newx)
}
\arguments{
\item{model}{: num2dif object}

\item{newx}{: data.frame with numerical variables}
}
\value{
data.frame with most important interacting features
}
\description{
Feature interaction:  f1 - f2
}
\examples{
  df = iris[1:100,]
  df$Species = as.factor(ifelse(df$Species == "setosa", 1, 0))
  m = num2dif(df[,1:4], df[,5])
  xdif = predict(m, df[,1:4]
  data = cbind(df, xdif)

}
