% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2div}
\alias{predict.num2div}
\title{Feature interaction:  f1 / f2}
\usage{
\method{predict}{num2div}(model, newx, smooth = 0.001)
}
\arguments{
\item{model}{: num2div object}

\item{newx}{: data.frame with numerical variables}
}
\value{
data.frame with most important interacting features
}
\description{
Feature interaction:  f1 / f2
}
\examples{
  df = iris[1:100,]
  df$Species = as.factor(ifelse(df$Species == "setosa", 1, 0))
  m = num2div(df[,1:4], df[,5])
  xdiv = predict(m, df[,1:4]
  data = cbind(df, xdiv)

}
