% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2dot}
\alias{predict.num2dot}
\title{Feature interaction:  f1 * f2}
\usage{
\method{predict}{num2dot}(model, newx)
}
\arguments{
\item{model}{: num2dot object}

\item{newx}{: data.frame with numerical variables}
}
\value{
data.frame with most important interacting features
}
\description{
Feature interaction:  f1 * f2
}
\examples{
  df = iris[1:100,]
  df$Species = as.factor(ifelse(df$Species == "setosa", 1, 0))
  m = num2dot(df[,1:4], df[,5])
  xdot = predict(m, df[,1:4]
  data = cbind(df, xdot)

}
