% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2dtf}
\alias{predict.num2dtf}
\title{Decision tree feature}
\usage{
\method{predict}{num2dtf}(model, newx)
}
\arguments{
\item{model}{: num2dtf object}

\item{newx}{: data.frame with numeric variables}
}
\value{
class label
}
\description{
Decision tree feature
}
\examples{
  m = num2dtf(Species ~ ., iris)
  dtf = predict(m, iris)
  cbind(iris, dtf)

}
