% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2gasvm}
\alias{predict.num2gasvm}
\title{Feature selection using the wrapper method}
\usage{
\method{predict}{num2gasvm}(model, newx, ...)
}
\arguments{
\item{model}{: num2gasvm object}

\item{newx}{: data.frame}
}
\value{
data.frame; the number of columns is reduced
}
\description{
Feature selection using the wrapper method
}
\examples{
  x = iris[1:100,1:4]
  y = ifelse(iris[1:100, 5] == "setosa", 1, 0)
  m = num2gasvm(x, y)
  newx = predict(m, x)#'

}
