% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2imp}
\alias{predict.num2imp}
\title{Feature selection using randomForest}
\usage{
\method{predict}{num2imp}(model, newx)
}
\arguments{
\item{model}{: num2imp object}

\item{newx}{: data.frame with numerical variables}
}
\value{
data.frame dimension may be reduced
}
\description{
Features are selected on the basis of variable$importance
}
\examples{
  m = num2imp(Species ~ ., iris)
  dat = predict(m, iris)

}
