% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2log}
\alias{predict.num2log}
\title{Identify which numericcal variables needs to be log
The decision is based on the measurement of their skewness}
\usage{
\method{predict}{num2log}(model, newx, smooth = 0.001)
}
\arguments{
\item{model}{: num2log object}

\item{newx}{: data.frame with numerical variables}

\item{smooth}{: double;  value is added to numerical variable before log:  log(newx[[i]] + smooth)}
}
\value{
data.frame; number of dimensions may be reduced
}
\description{
Identify which numericcal variables needs to be log
The decision is based on the measurement of their skewness
}
