% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2norm}
\alias{predict.num2norm}
\title{Continuous variable scaling}
\usage{
\method{predict}{num2norm}(model, newx, method = NULL, ...)
}
\arguments{
\item{model}{: num2norm object}

\item{newx}{: data.frame}
}
\value{
data.frame;  dim is preserved
}
\description{
Continuous variable scaling
}
\examples{
  m = num2norm(iris[,1:2])
  predict(m, iris[,1:4])

}
