% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2scale}
\alias{predict.num2scale}
\title{Feature scaling of numeric variables}
\usage{
\method{predict}{num2scale}(model, newx, method = "scaling")
}
\arguments{
\item{model}{: num2scale object}

\item{newx}{: data.frame}

\item{method}{: [scaling (default) | normalising | standardising ]
scaling = (x-xmin) / (xmax-xmin)
normalising = (x-xmean) / (xmax-xmin)
standardising = (x-xmean) / xsd}
}
\value{
data.frame;  dim is preserved
}
\description{
Feature scaling of numeric variables
}
\examples{
  m = num2scale(iris[,1:2])
  predict(m, iris[,1:4], )

}
