% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.num2trafo}
\alias{predict.num2trafo}
\title{Box-Cox transformation}
\usage{
\method{predict}{num2trafo}(model, newx, ...)
}
\arguments{
\item{model}{: num2trafo object}

\item{newx}{: data.frame}
}
\value{
data.frame;  dim is preserved
}
\description{
Box-Cox transformation
}
\examples{
  m = num2trafo(iris[,1:2])
  predict(m, iris[,1:4])

}
