% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{predict.nzv2rm}
\alias{predict.nzv2rm}
\title{Remove near zero variance (NZV) variables}
\usage{
\method{predict}{nzv2rm}(model, newx)
}
\arguments{
\item{model}{: nzv2rm object}

\item{newx}{: data.frame}
}
\value{
data.frame : number of columns is reduced
}
\description{
Remove near zero variance (NZV) variables
}
\examples{
  sel = rm_nzv(iris[,1:4])
  learn = iris[,sel]

}
