% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{rssvm}
\alias{rssvm}
\title{Building an ensemble of SVM classifiers using random sampling}
\usage{
rssvm(x, y, iter = 100, nsvm = 10)
}
\arguments{
\item{x}{: data.frame with numeric variables
nx = sqrt(ncol(x)) is used to randomly select variables per learning}

\item{y}{: vector with label [0,1]
n0 = sqrt(length(i0)) is used to randomly select negative class records}

\item{nsvm}{: integer; number of SVMs to be trained}
}
\value{
rsvm object
}
\description{
Building an ensemble of SVM classifiers using random sampling
}
\examples{
  dt = iris[1:100,]
  dt$Species = ifelse(dt$Species == "setosa", 1, 0)
  m = rsvm(dt[,1:4], dt$Species)
  predict(m, dt)

}
