% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{tfeat}
\alias{tfeat}
\title{Extract time related features}
\usage{
tfeat(x, width)
}
\arguments{
\item{x}{A vector with timestamps of type POSIXct}

\item{y}{A vector with group label}
}
\value{
A data.frame with features
}
\description{
Extract time related features
}
