% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{time2feat}
\alias{time2feat}
\title{Extract time related features}
\usage{
time2feat(x, width)
}
\arguments{
\item{x}{:  POSIXct containing YYYY-mm-dd}

\item{width}{: integer; window for rolling statistics}
}
\value{
data.frame
}
\description{
Extract time related features
}
\examples{
  

}
