% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{data2dt}
\alias{data2dt}
\title{Tree-based models and ensembles:  rpart::rpart}
\usage{
data2dt(f, x, ...)
}
\arguments{
\item{f}{: formula}

\item{x}{: data.frame}
}
\description{
Tree-based models and ensembles:  rpart::rpart
}
\examples{
  dt = rpart::rpart(f, data=x)
  p = predict(dt, newdata=iris, type="class")
  table(p, a=iris[["Species"]])

}
